fix the "ok" command response type (int -> double)

+++ lib/src/main/java/org/hibernate/omm/jdbc/MongodbStatement.java
@@ -36,7 +36,7 @@ public class MongodbStatement implements StatementAdapter {
         throwExceptionIfClosed();
         Document command = Document.parse(sql);
         Document commandResult = mongoDatabase.runCommand(command);
-        if (commandResult.getInteger("ok") != 1) {
+        if (commandResult.getDouble("ok") != 1.0) {
             throw new CommandRunFailSQLException();
         }
         return new MongodbResultSet(commandResult);
@@ -59,7 +59,7 @@ public class MongodbStatement implements StatementAdapter {
         Document command = Document.parse(sql);
         String collection = (String) command.entrySet().iterator().next().getValue();
         Document commandResult = mongoDatabase.runCommand(command);
-        if (commandResult.getInteger("ok") != 1) {
+        if (commandResult.getDouble("ok") != 1.0) {
             throw new CommandRunFailSQLException();
         }
         Document cursor = commandResult.get("cursor", Document.class);
@@ -121,7 +121,7 @@ public class MongodbStatement implements StatementAdapter {
                 moreResultsCommand.append("batchSize", fetchSize);
             }
             Document moreResults = mongoDatabase.runCommand(moreResultsCommand);
-            if (moreResults.getInteger("ok") != 1) {
+            if (moreResults.getDouble("ok") != 1.0) {
                 throw new CommandRunFailSQLException();
             }
             Document cursor = moreResults.get("cursor", Document.class);
